flask-backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
Authlib==1.3.2
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
convex==0.6.0
cryptography==43.0.1
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
grpcio==1.66.2
grpcio-tools==1.66.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.1
langchain-text-splitters==0.3.0
langsmith==0.1.123
llama_cpp_python==0.2.90
MarkupSafe==2.1.5
numpy==2.1.1
openai==1.50.2
orjson==3.10.7
packaging==24.1
portalocker==2.10.1
protobuf==5.28.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.11.3
requests==2.32.3
sniffio==1.3.1
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
validators==0.34.0
weaviate==0.1.0
weaviate-client==3.26.7
Werkzeug==3.0.4
